requirements.txt
altair==5.5.0
asn1crypto==1.5.1
attrs==25.3.0
blinker==1.9.0
boto3==1.38.41
botocore==1.38.41
cachetools==6.1.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cryptography==45.0.4
filelock==3.18.0
gitdb==4.0.12
GitPython==3.1.44
idna==3.10
Jinja2==3.1.6
jmespath==1.0.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
MarkupSafe==3.0.2
narwhals==1.43.1
numpy==2.3.1
packaging==25.0
pandas==2.3.0
pillow==11.2.1
platformdirs==4.3.8
protobuf==6.31.1
pyarrow==18.0.0
pycparser==2.22
pydeck==0.9.1
PyJWT==2.10.1
pyOpenSSL==25.1.0
python-dateutil==2.9.0.post0
pytz==2025.2
referencing==0.36.2
requests==2.32.4
rpds-py==0.25.1
s3transfer==0.13.0
six==1.17.0
smmap==5.0.2
snowflake-connector-python==3.15.0
sortedcontainers==2.4.0
streamlit==1.46.0
tenacity==9.1.2
toml==0.10.2
tomlkit==0.13.3
tornado==6.5.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
